moved soundSource to its own package in model
