+++ src/main/java/app/model/soundSource/SoundSourceType.java
@@ -1,0 +1,12 @@
+package app.model.soundSource;
+
+public enum SoundSourceType {
+    SIREN("siren");
+
+    public final String label;
+
+    SoundSourceType(String label)
+    {
+        this.label = label;
+    }
+}
